refactor: use urfave cli
