+++ go.mod
@@ -1,6 +1,8 @@
 module github.com/kijimaD/xruler
 
-go 1.21.3
+go 1.22
+
+toolchain go1.24.0
 
 require (
 	github.com/BurntSushi/xgb v0.0.0-20210121224620-deaf085860bc
@@ -10,4 +12,5 @@ require (
 require (
 	github.com/BurntSushi/freetype-go v0.0.0-20160129220410-b763ddbfe298 // indirect
 	github.com/BurntSushi/graphics-go v0.0.0-20160129215708-b43f31a4a966 // indirect
+	github.com/urfave/cli/v3 v3.4.1 // indirect
 )
